Move persistent volume constants and configurations away from model
