+++ shell/edit/persistentvolume/index.vue
@@ -13,9 +13,9 @@ import NodeAffinity from '@shell/components/form/NodeAffinity';
 import { Checkbox } from '@components/Form/Checkbox';
 import uniq from 'lodash/uniq';
 import UnitInput from '@shell/components/form/UnitInput';
+import { VOLUME_PLUGINS, LONGHORN_PLUGIN } from '@shell/config/persistentVolume';
 import { NODE, PVC, STORAGE_CLASS } from '@shell/config/types';
 import Loading from '@shell/components/Loading';
-import { LONGHORN_PLUGIN, VOLUME_PLUGINS } from '@shell/models/persistentvolume';
 import { _CREATE, _VIEW } from '@shell/config/query-params';
 import { clone } from '@shell/utils/object';
 import InfoBox from '@shell/components/InfoBox';
