+++ shell/edit/persistentvolume/plugins/longhorn.vue
@@ -3,7 +3,7 @@ import KeyValue from '@shell/components/form/KeyValue';
 import { LabeledInput } from '@components/Form/LabeledInput';
 import { RadioGroup } from '@components/Form/Radio';
 import { _CREATE } from '@shell/config/query-params';
-import { LONGHORN_DRIVER } from '@shell/models/persistentvolume';
+import { LONGHORN_DRIVER } from '@shell/config/types';
 
 export default {
   components: {
