+++ shell/edit/storage.k8s.io.storageclass/index.vue
@@ -11,8 +11,7 @@ import LabeledSelect from '@shell/components/form/LabeledSelect';
 import { _CREATE, _VIEW } from '@shell/config/query-params';
 import { PROVISIONER_OPTIONS } from '@shell/models/storage.k8s.io.storageclass';
 import { mapFeature, UNSUPPORTED_STORAGE_DRIVERS } from '@shell/store/features';
-import { CSI_DRIVER } from '@shell/config/types';
-import { LONGHORN_DRIVER } from '@shell/models/persistentvolume';
+import { CSI_DRIVER, LONGHORN_DRIVER } from '@shell/config/types';
 
 export default {
   name: 'StorageClass',
